+++ filter.php
@@ -53,12 +53,35 @@ class filter_synhi extends moodle_text_filter {
         if (is_string($text)) {
             if (($this->context->contextlevel >= CONTEXT_COURSE) && ($this->context->contextlevel <= CONTEXT_BLOCK)) {
                 // Do a quick check to see if we have a tag.
-                if ((strpos($text, '<pre') !== false) || (strpos($text, '<code') !== false)) {
-                    if (!self::$done) {
-                        $toolbox = \filter_synhi\toolbox::get_instance();
-                        $toolbox->highlight_page();
+                $prepos = strpos($text, '<pre');
+                $synpos = false;
+                if ($prepos !== false) {
+                    $synpos = $prepos + 4; // Number of characters in '<pre'.
+                } else {
+                    $codepos = strpos($text, '<code');
+                    if ($codepos !== false) {
+                        $synpos = $codepos + 5; // Number of characters in '<code'.
+                    }
+                }
+                if ($synpos !== false) {
+                    // Don't alter MathJax -> https://docs.moodle.org/310/en/MathJax_filter.
+                    if ((strpos($text, '$$') === false) &&
+                        (strpos($text, '[tex]') === false) &&
+                        (strpos($text, '<tex>') === false) &&
+                        // Above when SynHi filter is before MathJax / TeX fiters in admin list and below for below.
+                        (strpos($text, 'MathJax') === false)) {
 
-                        self::$done = true;
+                        $config = get_config('filter_synhi');
+                        if (!empty($config->engine)) {
+                            if ($config->engine == 'enlighterjs') {
+                                $text = substr($text, 0, $synpos).' class="synhi"'.substr($text, $synpos);
+                            }
+                            if (!self::$done) {
+                                $toolbox = \filter_synhi\toolbox::get_instance();
+                                $toolbox->highlight_page($config);
+                                self::$done = true;
+                            }
+                        }
                     }
                 }
             }